Little updated python wrapper tests
