+++ wrappers/java/src/test/java/org/hyperledger/indy/sdk/ledger/SubmitActionTest.java
@@ -5,7 +5,6 @@ import org.hyperledger.indy.sdk.did.Did;
 import org.hyperledger.indy.sdk.did.DidResults;
 import org.json.JSONObject;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
